requirements.txt
affine==2.4.0
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Deprecated==1.2.14
Django==5.0.6
django-cors-headers==4.4.0
django-geojson==4.1.0
django-leaflet==0.30.1
djangorestframework==3.15.2
djangorestframework-gis==1.0
Flask==3.0.3
folium==0.17.0
fonttools==4.53.1
geojsonio==0.0.3
geopandas==1.0.1
geos==0.2.3
gitdb==4.0.11
github3.py==4.0.1
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
ldap3==2.9.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==1.26.4
ogr==0.49.2
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
proj==0.2.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydirectory==0.1.15
pygame==2.6.0
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==4.7.0
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.10
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==70.2.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
sqlparse==0.5.0
tqdm==4.66.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xyzservices==2024.6.0
